Back-46 replace coverage type to BigDecimal

+++ hh-dep-monitoring-service/src/main/java/ru/hh/school/depmonitoring/dto/RepositoryDto.java
@@ -6,6 +6,7 @@ import java.io.Serializable;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Objects;
@@ -54,7 +55,7 @@ public class RepositoryDto implements Serializable {
     private List<ArtifactDto> artifacts;
 
     @Nullable
-    private Float coverage;
+    private BigDecimal coverage;
 
     @Nonnull
     public List<RepositoryLinkDto> getLinkUrls() {
@@ -169,11 +170,11 @@ public class RepositoryDto implements Serializable {
         this.artifacts = artifacts;
     }
 
-    public Float getCoverage() {
+    public BigDecimal getCoverage() {
         return coverage;
     }
 
-    public void setCoverage(Float coverage) {
+    public void setCoverage(BigDecimal coverage) {
         this.coverage = coverage;
     }
